+++ CHANGELOG.md
@@ -90,6 +90,7 @@
     - @vitejs/plugin-vue: ^6.0.3
     - @vitest/ui: ^4.0.16
     - @vitest/web-worker: ^4.0.16
+    - @vitest/eslint-plugin": ^1.6.6
     - cross-env: ^10.1.0
     - jsdom: ^26.0.0
     - jsdom-worker: ^0.3.0
@@ -121,8 +122,8 @@
   - dependencies:
     - url-parse  
   - devDependencies:
-    - zip-a-folder
     - css-loader
+    - eslint-plugin-mocha
     - file-loader
     - imports-loader
     - inquirer